backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
scikit-learn==1.5.0
scikit-image==0.24.0
pytesseract==0.3.13
ultralytics==8.4.21
joblib==1.4.2
Pillow==10.4.0
easyocr==1.7.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
keras-facenet==0.3.2
tensorflow==2.17.0

